move all Room use cases into separate controllers
