+++ src/controllers/CreatePropertyController.tsx
@@ -1,6 +1,6 @@
 ﻿import { supabase } from "../supabase/db";
 import { Database } from "../supabase/supabase";
-import { createRooms } from "./RoomController";
+import { createRooms } from "./CreateRoomController";
 
 
 export const createProperty = async (property: Database['public']['Tables']['Properties']['Insert'], rooms: string) => {
